Modified rbenv plugin to handle Homebrew-installed rbenv
The pull request from peterhoeg is great, except that it assumes the 'rbenv'
program is located at $HOME/.rbenv/bin -- which is NOT true if you installed
it with 'brew install rbenv'.

This change splits the "add .rbenv/bin to the PATH" action from the rest of
the rbenv initialization.